pyproject.toml
[project]
name = "py-superbot"
version = "0.1.0"
description = "Clean Architecture project in Python"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "python-telegram-bot[job-queue]>=20.0",
    "motor>=3.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
